fix(application): my application refactoring
