+++ src/components/features/applications/components/index.ts
@@ -1,2 +1,4 @@
 export { default as ApplicationForm } from './ApplicationForm'
 export { default as ApplicationModal } from './ApplicationModal'
+export { default as MyApplicationList } from './MyApplicationList'
+export { default as MyApplicationPageClient } from './MyApplicationPageClient'
